requirements.txt
grpcio
grpcio-tools
protobuf
scikit-learn
joblib